* commit '2e69567964e0ff7561e3404c2abf50c523a0d28d':
  build: resolve a compiler warning of -Wnarrowing

+++ common/stringutil.cpp
@@ -942,7 +942,7 @@ static const struct acl_role_name *find_acl_role(unsigned int rolenum)
 
 std::string AclRightsToString(unsigned int ror)
 {
-	switch (ror) {
+	switch (static_cast<int>(ror)) {
 	case -1:
 		return "Error reading ACL";
 	case -2:
